requirements.txt
# ─────────────────────────────────────────────────────────────
# TEXT-TO-SQL CHATBOT — Requirements
# ─────────────────────────────────────────────────────────────
# Install: pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────

# ─── LLM Providers ───────────────────────────────────────────
anthropic>=0.40.0           # Anthropic Claude
openai>=1.0.0               # OpenAI GPT + Gemini (OpenAI-compatible)
groq>=1.0.0                 # Groq (Llama, Mixtral)

# ─── API & Web ────────────────────────────────────────────────
fastapi>=0.100.0            # REST API framework
uvicorn>=0.20.0             # ASGI server
streamlit>=1.31.0           # Web UI
slowapi>=0.1.9              # Rate limiting for FastAPI

# ─── Database ─────────────────────────────────────────────────
sqlalchemy>=2.0.0           # SQL ORM & connection
psycopg2-binary>=2.9.0      # PostgreSQL driver

# ─── Vector Database ──────────────────────────────────────────
chromadb>=0.4.0             # Vector database for RAG

# ─── Data Processing ──────────────────────────────────────────
pandas>=2.0.0               # CSV loading & data manipulation
pydantic>=2.0.0             # Data validation & models

# ─── SQL & Config ─────────────────────────────────────────────
sqlparse>=0.4.0             # SQL parsing & validation
pyyaml>=6.0.0               # YAML config files
python-dotenv>=1.0.0        # .env file loading

# ─── Testing ──────────────────────────────────────────────────
pytest>=7.0.0               # Test framework
pytest-mock>=3.0.0          # Mocking for tests
Faker>=10.0.0               # Fake data generation